ref: extract logic to useNavigationPrompt hook
